+++ src/main/java/ru/mail/park/Application.java
@@ -7,9 +7,6 @@ import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
-/**
- * Created by Solovyev on 06/09/16.
- */
 @SpringBootApplication
 public class Application {
     public static void main(String[] args) {
