+++ src/main/java/ru/mail/park/controller/response/SuccessResponse.java
@@ -1,0 +1,14 @@
+package ru.mail.park.controller.response;
+
+public class SuccessResponse {
+    private String response;
+
+    public SuccessResponse(String response) {
+        this.response = response;
+    }
+
+    @SuppressWarnings("unused")
+    public String getResponse() {
+        return response;
+    }
+}
