--- src/main/resources/hibernate.cfg.xml
@@ -1,11 +1,0 @@
-<?xml version='1.0' encoding='utf-8'?>
-<!DOCTYPE hibernate-configuration PUBLIC
-        "-//Hibernate/Hibernate Configuration DTD//EN"
-        "http://www.hibernate.org/dtd/hibernate-configuration-3.0.dtd">
-<hibernate-configuration>
-    <session-factory>
-        <property name="show_sql">true</property>
-        <!-- NOTE that after automapping with IDEA property username and password will delete -->
-        <mapping class="ru.mail.park.entities.UserProfileEntity"/>
-    </session-factory>
-</hibernate-configuration>
